src/requirements.txt
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
numpy==1.26.3
tqdm==4.67.1
hydra-core==1.3.2
iopath==0.1.10
pillow==10.2.0
omegaconf==2.3.0
antlr4-python3-runtime==4.9.3
portalocker==3.0.0
PyYAML==6.0.2
packaging==24.2
tifffile==2024.9.20
matplotlib==3.9.2
scipy==1.14.1
scikit-image==0.24.0
imagecodecs==2024.9.22
torchio==0.20.2  
monai==1.4.0
prefetch_generator==1.0.3
opencv-python-headless==4.10.0
zarr==2.18.2